+++ app/src/main/java/org/simple/clinic/overdue/download/formatdialog/SelectOverdueDownloadFormatDialog.kt
@@ -8,6 +8,7 @@ import android.view.View
 import android.view.ViewGroup
 import com.jakewharton.rxbinding3.view.clicks
 import com.jakewharton.rxbinding3.widget.checkedChanges
+import com.spotify.mobius.Init
 import com.spotify.mobius.functions.Consumer
 import io.reactivex.Observable
 import io.reactivex.rxkotlin.cast
@@ -72,6 +73,8 @@ class SelectOverdueDownloadFormatDialog : BaseDialog<
   override fun createUpdate() = SelectOverdueDownloadFormatUpdate()
 
   override fun defaultModel() = SelectOverdueDownloadFormatModel.create(openAs)
+
+  override fun createInit() = SelectOverdueDownloadFormatInit()
 
   override fun events() = Observable
       .merge(