+++ app/src/main/java/org/simple/clinic/overdue/download/formatdialog/SelectOverdueDownloadFormatUiRenderer.kt
@@ -32,6 +32,8 @@ class SelectOverdueDownloadFormatUiRenderer(
     when (model.openAs) {
       Download -> renderDownloadUi()
       Share -> renderShareUi()
+      SharingInProgress -> { // do nothing
+      }
     }.exhaustive()
   }