+++ app/src/test/java/org/simple/clinic/home/overdue/OverdueUpdateTest.kt
@@ -11,6 +11,7 @@ import org.simple.clinic.TestData
 import org.simple.clinic.analytics.NetworkConnectivityStatus.ACTIVE
 import org.simple.clinic.analytics.NetworkConnectivityStatus.INACTIVE
 import org.simple.clinic.facility.FacilityConfig
+import org.simple.clinic.overdue.download.OverdueListFileFormat.CSV
 import java.time.LocalDate
 import java.util.Optional
 import java.util.UUID
@@ -18,7 +19,7 @@ import java.util.UUID
 class OverdueUpdateTest {
 
   private val dateOnClock = LocalDate.parse("2018-01-01")
-  private val updateSpec = UpdateSpec(OverdueUpdate(dateOnClock))
+  private val updateSpec = UpdateSpec(OverdueUpdate(date = dateOnClock, canGeneratePdf = true))
   private val defaultModel = OverdueModel.create()
 
   @Test
@@ -61,7 +62,6 @@ class OverdueUpdateTest {
 
   @Test
   fun `when current facility is loaded and overdue list changes feature is enabled, then load overdue appointments with patients with no phone numbers`() {
-    val updateSpec = UpdateSpec(OverdueUpdate(dateOnClock))
     val facility = TestData.facility(
         uuid = UUID.fromString("6d66fda7-7ca6-4431-ac3b-b570f1123624"),
         facilityConfig = FacilityConfig(
@@ -91,13 +91,26 @@ class OverdueUpdateTest {
   }
 
   @Test
-  fun `when download overdue list button is clicked and network is connected, then open select download format dialog`() {
+  fun `when download overdue list button is clicked, network is connected and pdf can be generated, then open select download format dialog`() {
     updateSpec
         .given(defaultModel)
         .whenEvent(DownloadOverdueListClicked(networkStatus = Optional.of(ACTIVE)))
         .then(assertThatNext(
             hasNoModel(),
             hasEffects(OpenSelectDownloadFormatDialog)
+        ))
+  }
+
+  @Test
+  fun `when download overdue list button is clicked, network is connected and pdf can not be generated, then schedule download`() {
+    val updateSpec = UpdateSpec(OverdueUpdate(date = dateOnClock, canGeneratePdf = false))
+
+    updateSpec
+        .given(defaultModel)
+        .whenEvent(DownloadOverdueListClicked(networkStatus = Optional.of(ACTIVE)))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(ScheduleDownload(CSV))
         ))
   }
 
@@ -113,13 +126,26 @@ class OverdueUpdateTest {
   }
 
   @Test
-  fun `when share overdue list button is clicked and network is connected, then open select share format dialog`() {
+  fun `when share overdue list button is clicked, network is connected and pdf can be generated, then open select share format dialog`() {
     updateSpec
         .given(defaultModel)
         .whenEvent(ShareOverdueListClicked(networkStatus = Optional.of(ACTIVE)))
         .then(assertThatNext(
             hasNoModel(),
             hasEffects(OpenSelectShareFormatDialog)
+        ))
+  }
+
+  @Test
+  fun `when share overdue list button is clicked, network is connected but pdf can not be generated, then open progress for share dialog`() {
+    val updateSpec = UpdateSpec(OverdueUpdate(date = dateOnClock, canGeneratePdf = false))
+
+    updateSpec
+        .given(defaultModel)
+        .whenEvent(ShareOverdueListClicked(networkStatus = Optional.of(ACTIVE)))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(OpenSharingInProgressDialog)
         ))
   }
 }